Allow for wands that use a custom skull texture as their icon

+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/FreezeAction.java
@@ -73,7 +73,7 @@ public class FreezeAction extends BaseSpellAction implements BlockAction
 		{
 			if (block.getData() < 7)
 			{
-				applyMaterial.setData((byte)(block.getData() + 1));
+				applyMaterial.setData((short)(block.getData() + 1));
 			}
 		}
 		applyMaterial.modify(block);
